fix: move types and validations to model
