+++ server/src/app/shared/model/types.ts
